refactor: move DashboardFilterBar to filters folder (#212)
