+++ frontend/src/pages/inventory/InventoryPage.tsx
@@ -41,7 +41,7 @@ import { Doughnut, Line } from 'react-chartjs-2'
 import { format } from 'date-fns'
 import { useNavigate } from 'react-router-dom'
 import PageHeader from '@/components/common/PageHeader'
-import { DashboardFilterBar } from '@/components/dashboard/DashboardFilterBar'
+import { DashboardFilterBar } from '@/components/filters/DashboardFilterBar'
 import { useDashboardFilters } from '@/hooks/useDashboardFilters'
 import { useInventoryAnalytics } from './hooks/useInventoryAnalytics'
 import { formatCurrency } from '@/utils/formatters'
